Rename "bodo.submit" to "bodo.spawn" (#93)
