+++ BodoSQL/bodosql/context.py
@@ -1032,7 +1032,7 @@ class BodoSQLContext:
 
     def sql(self, sql, params_dict=None, dynamic_params_list=None, **jit_options):
         import bodosql
-        from bodo.submit.spawner import SubmitDispatcher
+        from bodo.spawn.spawner import SpawnDispatcher
 
         if params_dict is None:
             params_dict = {}
@@ -1099,11 +1099,11 @@ class BodoSQLContext:
 
             dispatcher = bodo.jit(impl, **jit_options)
 
-            # Save BodoSQL globals in SubmitDispatcher to be handled in pickling
+            # Save BodoSQL globals in SpawnDispatcher to be handled in pickling
             # properly. Internal CASE implementation strings may use some globals that
             # are not visible to cloudpickle as used by the function. See:
             # test_json_fns.py::test_object_construct_keep_null[no_nested-no_null-with_case]
-            if isinstance(dispatcher, SubmitDispatcher):
+            if isinstance(dispatcher, SpawnDispatcher):
                 # __builtins__ which is added to glbls by exec causes issues in
                 # Jupyter/IPython pickling.
                 glbls.pop("__builtins__", None)
