+++ bodo/pandas/array_manager.py
@@ -21,7 +21,7 @@ except ModuleNotFoundError:
 
 import bodo.user_logging
 from bodo.pandas.lazy_metadata import LazyMetadataMixin
-from bodo.submit.utils import debug_msg
+from bodo.spawn.utils import debug_msg
 
 
 class LazyArrayManager(ArrayManager, LazyMetadataMixin[ArrayManager]):
